refactor: SM-187 DetailModal CocktailList 내부로 이동
