+++ src/components/domain/CocktailList/types.ts
@@ -5,7 +5,6 @@ import type { ICocktailSimple } from '@models/types';
 interface CocktailListProps extends HTMLAttributes<HTMLDivElement> {
   cocktailList: ICocktailSimple[];
   noResultMsg?: string;
-  handleAlbumClick?(id: number): void;
 }
 
 interface StyledMotionWrapperProps {
